renamed:    src/main/java/com/dusinski/holidaycalendar/controllers/MainController.java -> src/main/java/com/dusinski/holidaycalendar/controller/MainController.java 	new file:   src/main/java/com/dusinski/holidaycalendar/controller/UserController.java 	renamed:    src/main/java/com/dusinski/holidaycalendar/entities/CheckStartEndDate.java -> src/main/java/com/dusinski/holidaycalendar/customvalidator/CheckStartEndDate.java 	renamed:    src/main/java/com/dusinski/holidaycalendar/entities/StartEndDateValidator.java -> src/main/java/com/dusinski/holidaycalendar/customvalidator/StartEndDateValidator.java 	deleted:    src/main/java/com/dusinski/holidaycalendar/entities/User.java 	renamed:    src/main/java/com/dusinski/holidaycalendar/entities/CalendarEvent.java -> src/main/java/com/dusinski/holidaycalendar/model/CalendarEvent.java 	renamed:    src/main/java/com/dusinski/holidaycalendar/entities/EventConfirmationToken.java -> src/main/java/com/dusinski/holidaycalendar/model/EventConfirmationToken.java 	new file:   src/main/java/com/dusinski/holidaycalendar/model/User.java 	renamed:    src/main/java/com/dusinski/holidaycalendar/service/CalendarEventRepository.java -> src/main/java/com/dusinski/holidaycalendar/repository/CalendarEventRepository.java 	renamed:    src/main/java/com/dusinski/holidaycalendar/service/EventConfirmationTokenRepository.java -> src/main/java/com/dusinski/holidaycalendar/repository/EventConfirmationTokenRepository.java 	renamed:    src/main/java/com/dusinski/holidaycalendar/service/UserRepository.java -> src/main/java/com/dusinski/holidaycalendar/repository/UserRepository.java 	renamed:    src/main/java/com/dusinski/holidaycalendar/service/EmailSenderService.java -> src/main/java/com/dusinski/holidaycalendar/services/EmailSenderService.java 	new file:   src/main/java/com/dusinski/holidaycalendar/services/UserService.java 	modified:   src/main/resources/application.properties 	modified:   src/main/resources/input_data.sql 	new file:   src/main/resources/templates/user/showUsers.html
