+++ src/main/java/com/dusinski/holidaycalendar/controller/MainController.java
@@ -1,13 +1,13 @@
 
-package com.dusinski.holidaycalendar.controllers;
+package com.dusinski.holidaycalendar.controller;
 
-import com.dusinski.holidaycalendar.entities.CalendarEvent;
-import com.dusinski.holidaycalendar.entities.EventConfirmationToken;
-import com.dusinski.holidaycalendar.entities.User;
-import com.dusinski.holidaycalendar.service.CalendarEventRepository;
-import com.dusinski.holidaycalendar.service.EmailSenderService;
-import com.dusinski.holidaycalendar.service.EventConfirmationTokenRepository;
-import com.dusinski.holidaycalendar.service.UserRepository;
+import com.dusinski.holidaycalendar.model.CalendarEvent;
+import com.dusinski.holidaycalendar.model.EventConfirmationToken;
+import com.dusinski.holidaycalendar.model.User;
+import com.dusinski.holidaycalendar.repository.CalendarEventRepository;
+import com.dusinski.holidaycalendar.services.EmailSenderService;
+import com.dusinski.holidaycalendar.repository.EventConfirmationTokenRepository;
+import com.dusinski.holidaycalendar.repository.UserRepository;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,15 +19,13 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 
 
 @Controller
 @RequestMapping
 public class MainController {
 
-    @Autowired
-    private UserRepository userRepository;
+
 
     @Autowired
     private CalendarEventRepository calendarEventRepository;
@@ -41,45 +39,45 @@ public class MainController {
     @Autowired
     private EmailSenderService emailSenderService;
 
-    @RequestMapping(path = "/add") // Map ONLY POST Requests
-    public @ResponseBody
-    String addNewUser(@RequestParam("name") String name
-            , @RequestParam("email") String email) {
-        // @ResponseBody means the returned String is the response, not a view name
-        // @RequestParam means it is a parameter from the GET or POST request
+//    @RequestMapping(path = "/add") // Map ONLY POST Requests
+//    public @ResponseBody
+//    String addNewUser(@RequestParam("name") String name
+//            , @RequestParam("email") String email) {
+//        // @ResponseBody means the returned String is the response, not a view name
+//        // @RequestParam means it is a parameter from the GET or POST request
+//
+//        User n = new User(name, email);
+//        userRepository.save(n);
+//        return "Saved";
+//    }
 
-        User n = new User(name, email);
-        userRepository.save(n);
-        return "Saved";
-    }
+//    @GetMapping(path = "/all")
+//    public @ResponseBody
+//    Iterable<User> getAllUsers() {
+//
+//        return userRepository.findAll();
+//    }
 
-    @GetMapping(path = "/all")
-    public @ResponseBody
-    Iterable<User> getAllUsers() {
-
-        return userRepository.findAll();
-    }
-
-    @GetMapping(path = "/test")
-    public @ResponseBody
-    String testResponse() {
-
-        return "Test String";
-    }
-
-
-    @GetMapping("/greeting")
-    public String greetingForm(Model model) {
-        model.addAttribute("user", new User());
-        return "greeting";
-    }
-
-    @PostMapping("/greeting")
-    public String greetingSubmit(@ModelAttribute User user, Model model) {
-    model.addAttribute("user", user);
-        userRepository.save(user);
-    return "result";
-    }
+//    @GetMapping(path = "/test")
+//    public @ResponseBody
+//    String testResponse() {
+//
+//        return "Test String";
+//    }
+//
+//
+//    @GetMapping("/greeting")
+//    public String greetingForm(Model model) {
+//        model.addAttribute("user", new User());
+//        return "greeting";
+//    }
+//
+//    @PostMapping("/greeting")
+//    public String greetingSubmit(@ModelAttribute User user, Model model) {
+//    model.addAttribute("user", user);
+//        userRepository.save(user);
+//    return "result";
+//    }
 
     @GetMapping(path="/")
     public String calendarForm( Model model) throws JsonProcessingException {
