--- src/main/java/com/dusinski/holidaycalendar/entities/User.java
@@ -1,60 +1,0 @@
-//entites
-
-package com.dusinski.holidaycalendar.entities;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Entity
-public class User {
-    @Id
-    @GeneratedValue(strategy= GenerationType.IDENTITY)
-    private long id;
-//    private String name;
-    private String email;
-    public long getId() {
-        return id;
-    }
-    public User(){
-//        this.email="test_name";
-//        this.email="test_email";
-    }
-
-    public User(String name, String email){
-//        this.name=name;
-        this.email=email;
-    }
-
-//    @Override
-//    public String toString(){
-//        return String.format("Customer[id=%d], firstname='%s'", id, name,email);
-//
-//    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-//    public String getName() {
-//        return name;
-//    }
-
-//    public void setName(String name) {
-//        this.name = name;
-//    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-
-
-
-
-}
