+++ src/main/resources/application.properties
@@ -43,6 +43,9 @@ spring.mail.properties.mail.smtp.writetimeout= 5000
 # start, end date validation
 #
 # token confirmation of new event
+
+# when delete user delete also his events
+
 # admin form - see all events, could create user
 #  user class - creation of new user, user could be admin, user could belong to manager
 # logging daemon
