+++ src/main/resources/input_data.sql
@@ -1,5 +1,5 @@
 insert into Calendar_Event(event_id,title, start, end, user_Id, is_Enabled) values (default ,'event1','2020-08-21','2020-08-23',1, true );
 insert into Calendar_Event(event_id,title, start, end, user_Id, is_Enabled) values (default,'event1', '2020-08-21','2020-08-25',1, true);
 
-insert into User(id,email) values (default,'test@test.pl');
+insert into User(id,email, is_admin) values (default,'dominik.dusinski@gmail.com', true);
