+++ src/main/resources/templates/user/showUsers.html
@@ -1,0 +1,88 @@
+<!doctype html>
+<html xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset='utf-8'/>
+
+    <!-- Bootstrap CSS -->
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+    <style>
+        html, body {
+            overflow: hidden; /* don't do scrollbars */
+            font-family: Arial, Helvetica Neue, Helvetica, sans-serif;
+            font-size: 14px;
+            padding-top: 20px;
+        }
+    </style>
+
+    <title>Holiday Calendar Booking Engine</title>
+
+</head>
+<body>
+<nav class="navbar navbar-expand-md navbar-dark bg-dark fixed-top">
+    <div class="collapse navbar-collapse" id="navbarsExampleDefault">
+        <ul class="navbar-nav mr-auto">
+            <li class="nav-item ">
+                <a class="nav-link" href="/">Calendar view</a>
+            </li>
+            <li class="nav-item ">
+                <a class="nav-link" href="/show">Show my holidays</a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" href="/addEvent">Add new holiday</a>
+            </li>
+            <li class="nav-item active">
+                <a class="nav-link" href="/user/show">Show list of users</a>
+            </li>
+        </ul>
+    </div>
+</nav>
+
+
+
+<main role="main" class="container">
+
+    <div class="starter-template">
+        <p class="lead">
+        <h1>Holiday Calendar - List of all users</h1>
+        </p>
+    </div>
+
+</main><!-- /.container -->
+<div th:switch="${userList}">
+    <h2 th:case="null">No Events yet!</h2>
+    <div th:case="*">
+        <h2>Events</h2>
+        <table class="table table-striped">
+            <thead>
+            <tr>
+                <th scope="col">User id</th>
+                <th scope="col">email</th>
+                <th scope="col">is Admin</th>
+                <th scope="col">Click to delete</th>
+            </tr>
+            </thead>
+            <tbody>
+            <tr th:each="user : ${userList}">
+                <td scope="row" th:text="${user.id}"></td>
+                <td th:text="${user.email}"></td>
+                <td th:text="${user.admin}"></td>
+
+                <td><a th:href="@{/user/delete/{id}(id=${user.id})}" class="btn btn-danger" onclick="return confirm('Are you sure to delete User? His holiday events will be erased too')">Delete User</a> </td>
+            </tr>
+            </tbody>
+        </table>
+    </div>
+
+</div>
+
+
+
+<!-- Bootstrap core JavaScript
+================================================== -->
+<!-- Placed at the end of the document so the pages load faster -->
+<script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js" integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q" crossorigin="anonymous"></script>
+<script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js" integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl" crossorigin="anonymous"></script>
+</body>
+</html>
